src/third_party/blink/web_tests/svg/dom/SVGViewSpec-expected.txt

CONSOLE ERROR: line 88: Uncaught SyntaxError: Unexpected end of input
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
